Upgrade to jquery 3.6.0
