+++ funnel/assets/cypress/integration/05_publishProject.js
@@ -1,10 +1,10 @@
 /* eslint-disable global-require */
-describe('Verify roles of editor', () => {
+describe('Verify roles of editor and publish project', () => {
   const { editor } = require('../fixtures/user.json');
   const profile = require('../fixtures/profile.json');
   const project = require('../fixtures/project.json');
 
-  it('Access available for editor in project settings', () => {
+  it('Verify roles of editor and publish project', () => {
     cy.login(`/${profile.title}`, editor.username, editor.password);
     cy.get(`[data-cy-title="${project.title}"]`).first().click();
     cy.location('pathname').should('contain', project.url);
@@ -20,5 +20,12 @@ describe('Verify roles of editor', () => {
     cy.get('a[data-cy="manage-labels"]').should('exist');
     cy.get('a[data-cy="setup-ticket-events"]').should('not.exist');
     cy.get('a[data-cy="scan-checkin"]').should('not.exist');
+    cy.get('button[data-cy-state="publish"]').click();
+    cy.location('pathname').should('contain', project.url);
+    cy.get('a[data-cy="project-menu"]:visible').click();
+    cy.wait(1000);
+    cy.get('a[data-cy-navbar="settings"]:visible').click();
+    cy.location('pathname').should('contain', 'settings');
+    cy.get('[data-cy="project-state"]').contains('Published');
   });
 });
