+++ funnel/assets/cypress/integration/06_manageVenue.js
