--- funnel/assets/cypress/integration/06_publishProject.js
@@ -1,23 +1,0 @@
-/* eslint-disable global-require */
-describe('Publish project', () => {
-  const { editor } = require('../fixtures/user.json');
-  const profile = require('../fixtures/profile.json');
-  const project = require('../fixtures/project.json');
-
-  it('Publish project', () => {
-    cy.login(`/${profile.title}`, editor.username, editor.password);
-    cy.get(`[data-cy-title="${project.title}"]`).first().click();
-    cy.location('pathname').should('contain', project.url);
-    cy.get('a[data-cy="project-menu"]:visible').click();
-    cy.wait(1000);
-    cy.get('a[data-cy-navbar="settings"]:visible').click();
-    cy.location('pathname').should('contain', 'settings');
-    cy.get('button[data-cy-state="publish"]').click();
-    cy.location('pathname').should('contain', project.url);
-    cy.get('a[data-cy="project-menu"]:visible').click();
-    cy.wait(1000);
-    cy.get('a[data-cy-navbar="settings"]:visible').click();
-    cy.location('pathname').should('contain', 'settings');
-    cy.get('[data-cy="project-state"]').contains('Published');
-  });
-});
