+++ funnel/assets/cypress/integration/07_addCFP.js
@@ -1,11 +1,13 @@
 /* eslint-disable global-require */
-describe('Add labels to project', () => {
+describe('Add CFP and labels to project', () => {
   const { editor } = require('../fixtures/user.json');
+  const cfp = require('../fixtures/cfp.json');
   const profile = require('../fixtures/profile.json');
   const project = require('../fixtures/project.json');
+  const dayjs = require('dayjs');
   const labels = require('../fixtures/labels.json');
 
-  it('Add labels', () => {
+  it('Add CFP and labels', () => {
     cy.login(
       `/${profile.title}/${project.url}`,
       editor.username,
@@ -16,6 +18,28 @@ describe('Add labels to project', () => {
     cy.wait(1000);
     cy.get('a[data-cy-navbar="settings"]:visible').click();
     cy.location('pathname').should('contain', 'settings');
+    cy.get('a[data-cy="add-cfp"]').click();
+    cy.location('pathname').should('contain', '/cfp');
+
+    cy.get('#field-instructions')
+      .find('.CodeMirror textarea')
+      .type(cfp.instructions, { force: true });
+    cy.get('button[name="open-now"]').click();
+    const cfpEndDay = dayjs().add(20, 'days').format('YYYY-MM-DDTHH:mm');
+    cy.get('#cfp_end_at').type(cfpEndDay);
+    cy.get('button[data-cy="add-cfp"]').click();
+    cy.location('pathname').should('contain', project.url);
+    cy.get('a[data-cy="project-menu"]:visible').click();
+    cy.wait(1000);
+    cy.get('a[data-cy-navbar="settings"]:visible').click();
+    cy.location('pathname').should('contain', 'settings');
+    cy.get('button[data-cy-cfp=open_cfp]').click();
+    cy.location('pathname').should('contain', project.url);
+    cy.get('a[data-cy="project-menu"]:visible').click();
+    cy.wait(1000);
+    cy.get('a[data-cy-navbar="settings"]:visible').click();
+    cy.location('pathname').should('contain', 'settings');
+    cy.get('[data-cy="cfp-state"]').contains('Open');
     cy.get('a[data-cy="manage-labels"').click();
     cy.location('pathname').should('contain', '/labels');
 
@@ -101,7 +125,7 @@ describe('Add labels to project', () => {
       .trigger('mouseover', { which: 1, force: true, view: window })
       .trigger('mousedown', { which: 1, force: true, view: window })
       .trigger('mousemove', {
-        pageX: 500,
+        pageX: 650,
         pageY: 550,
         force: true,
         view: window,
