+++ funnel/assets/cypress/integration/11_respondToAttend.js
