+++ funnel/assets/cypress/integration/12_verifyPromotorRole.js
