+++ funnel/assets/cypress/integration/13_setupEvent.js
