+++ funnel/assets/cypress/integration/15_checkin.js
