+++ funnel/assets/cypress/integration/17_badge.js
