+++ funnel/assets/cypress/integration/20_verifyattendeeList.js
