+++ funnel/assets/cypress/integration/21_viewSchedule.js
