+++ funnel/assets/cypress/integration/22_addSessionVideo.js
