+++ funnel/assets/cypress/integration/23_viewVideos.js
