+++ funnel/assets/cypress/integration/24_removeSessionVideo.js
@@ -1,6 +1,6 @@
 /* eslint-disable global-require */
 describe('Remove video to session', () => {
-  const { editor } = require('../fixtures/user.json');
+  const { admin } = require('../fixtures/user.json');
   const project = require('../fixtures/project.json');
   const proposal = require('../fixtures/proposal.json');
 
@@ -8,7 +8,7 @@ describe('Remove video to session', () => {
     cy.server();
     cy.route('**/viewsession-popup').as('view-session');
 
-    cy.login('/', editor.username, editor.password);
+    cy.login('/', admin.username, admin.password);
 
     cy.get('.upcoming')
       .find('.card--upcoming')
