+++ funnel/assets/cypress/integration/26_search.js
