+++ funnel/templates/layout.html.jinja2
@@ -205,17 +205,10 @@
   {% block basefooter -%}
   {%- endblock %}
 
-  {% block jquery -%}
-    <script src="//ajax.googleapis.com/ajax/libs/jquery/2.2.4/jquery.min.js"></script>
-    <script type="text/javascript">
-      if (typeof jQuery === 'undefined') {
-        var jqueryScript = document.createElement('script');
-        jqueryScript.setAttribute('type','text/javascript');
-        jqueryScript.setAttribute('src','{{ url_for("baseframe.static", filename="js/jquery-2.2.4.min.js") }}');
-        document.head.appendChild(jqueryScript);
-      }
-    </script>
-  {%- endblock %}
+  <script src="//ajax.googleapis.com/ajax/libs/jquery/3.6.0/jquery.min.js"></script>
+  <script type="text/javascript">
+    window.jQuery || document.write('<script type="text/javascript" src="{{ url_for("baseframe.static", filename="js/jquery-3.6.0.min.js") }}"><\/script>');
+  </script>
 
   {%- for asset in config.get('ext_js', []) %}
     <script type="text/javascript" src="{{ asset|ext_asset_url }}"></script>
